removing expirationdate class and changing to apidate to make the use and gson interceptors of this class more generic
