+++ src/main/java/br/com/moip/request/ApiDateRequest.java
@@ -1,0 +1,17 @@
+package br.com.moip.request;
+
+import java.util.Date;
+
+public class ApiDateRequest {
+    private Date date;
+
+    public Date getDate() {
+        return date;
+    }
+
+    public ApiDateRequest date(Date date) {
+        this.date = date;
+
+        return this;
+    }
+}
